feat(templates): drop glbmodel editor, keep read-only glb viewer
Removes the interactive 3D editor tier and the Edit bridge, leaving only the
read-only glb/gltf viewer:

- glb/template.html: remove the Edit button + the project-scaffold/import
  bridge; the file is now a pure read-only viewer.
- delete the glbmodel/ template (template.html, project_ops.py, sculpt_bake.py,
  import_part.mjs, placements.mjs, sculpt_bake.mjs) and its .glbproj/ registry
  entry.
- vendor: drop gltf-transform.bundle.mjs + its build entry/dep (editor-only),
  and drop TransformControls from three-entry.mjs (editor-only); rebuild
  three.bundle.mjs (742KB -> 704KB).
- tests: replace test_glbmodel_project.py with test_glb_template.py covering
  glb/gltf registry resolution (and that .glbproj no longer routes to an editor).
