--- scripts/vendor-three/gltf-transform-entry.mjs
@@ -1,4 +1,0 @@
-// Entry for gltf-transform.bundle.mjs — the @gltf-transform/core surface the
-// glbmodel editor's client-side import (import_part.mjs) needs, re-exported as
-// one self-contained ESM bundle for /template-assets/ (no CDN at runtime).
-export { Document, WebIO, VertexLayout } from '@gltf-transform/core';
